tambah form login

+++ form/list_user.php
@@ -2,20 +2,32 @@
     if (!isset($_SESSION['nama'])) {
         header('location: ../index.php'); // Redirect to the login page if not logged in
         exit(); }
+    include '../config/koneksi.php';
+    $level = $_SESSION['level'];
+    $edit_hapus_visible = false;
+
+    if ($level == 1) {
+        $edit_hapus_visible = true;
+    } elseif ($level == 2) {
+        $edit_hapus_visible = true;
+    } elseif ($level == 3) {
+        $edit_hapus_visible = true;
+    }
 ?>
 <!DOCTYPE html>
 <html>
 <head>
     <title>List User</title>
+    <link rel="stylesheet" href="../Config/table.css">
 </head>
 <body>
     <h2>Pemograman 1 2023</h2>
     <br>
-    <a href="../Home.html">Home</a>
+    <?php if ($edit_hapus_visible): ?>
+    <a href="?page=luser">+ Tambah User</a>
+    <?php endif; ?>
     <br>
-    <a href="tambah_user.php">+ Tambah User</a>
-    <br>
-    <table border="1">
+    <table border="1" class="styled-table">
         <tr>
             <th>Id</th>
             <th>Nama</th>
@@ -35,10 +47,16 @@
             <td><?php echo $d['password'];?></td>
             <td><?php echo $d['level'];?></td>
             <td><?php echo $d['status'];?></td>
+            <?php if ($edit_hapus_visible): ?>
             <td>
+            <?php if ($level == 1 || ($level == 2 && ($d['level'] == 3 || $d['level'] == 4)) || ($level == 3 && $d['level'] == 4)): ?>
                 <a href="edit_user.php?id=<?php echo $d['id_user']; ?>">Edit</a>
+                <?php endif; ?>
+                <?php if ($level == 1 || ($level == 2 && ($d['level'] == 3 || $d['level'] == 4)) || ($level == 3 && $d['level'] == 4)): ?>
                 <a href="hapus_user.php?id=<?php echo $d['id_user']; ?>">Hapus</a>
+                <?php endif; ?>
             </td>
+            <?php endif; ?>
         </tr>
         <?php
             }
